[formatter] Extract TypeFormatter and ListFormatter interfaces
TypeFormatter depends on ListFormatter, so we have to extract them simultaneously
