+++ core/src/main/kotlin/org/jetbrains/ktfmt/format/KotlinLangInputAstVisitor.kt
@@ -1,10 +1,9 @@
 package org.jetbrains.ktfmt.format
 
 import com.google.googlejavaformat.OpsBuilder
+import org.jetbrains.ktfmt.format.visitor.KotlinLangListFormatter
 
 internal class KotlinLangInputAstVisitor(
     options: FormattingOptions,
     builder: OpsBuilder,
-) : KotlinInputAstVisitor(options, builder) {
-  override val forceAnnotationBreaks: Boolean = true
-}
+) : KotlinInputAstVisitor(options, builder), KotlinLangListFormatter
